+++ watcher/github-pr/pkg/factory/factory.go
@@ -59,6 +59,8 @@ func CreateWatcher(
 	taskCreationFilter filter.TaskCreationFilter,
 	startTime libtime.DateTime,
 	trustedAuthors []string,
+	maxSlugLen int,
+	maxTitleLen int,
 ) (pkg.Watcher, func(), error) {
 	branch := base.Branch(stage)
 	createSender, updateFrontmatterSender, cleanup, err := CreateKafkaSenders(ctx, brokers, branch)
@@ -80,6 +82,8 @@ func CreateWatcher(
 		stage,
 		pkg.NewMetrics(),
 		trustDecision,
+		maxSlugLen,
+		maxTitleLen,
 	)
 	return w, cleanup, nil
 }